Added Contour shaders
Two shaders, Contour and Contour Outline, were added to a new Shaders
folder. They are experimental shaders that may be used in the future.
In order to use them, create a new material and choose “Custom/Contour”
from the shader dropdown menu. Tangent Color is the color that a
surface tangent to the viewing direction will take, and normal color is
the color for the portion of surfaces that are normal to the view
direction.